Compile translations at run time
