+++ .gitignore
@@ -3,3 +3,4 @@ rls
 /target
 **/*.rs.bk
 rls
+translations
